+++ src/common/constants/api.js
@@ -1,1 +1,3 @@
 export const TITLES_API = 'orders/titles';
+export const TITLE_API = `${TITLES_API}/:{id}`;
+export const PO_LINES_API = 'orders/order-lines';