+++ src/common/resources/titleResources.js
@@ -2,10 +2,18 @@ import {
   baseManifest,
 } from '@folio/stripes-acq-components';
 
-import { TITLES_API } from '../constants';
+import {
+  TITLE_API,
+  TITLES_API,
+} from '../constants';
 
 export const titlesResource = {
   ...baseManifest,
   path: TITLES_API,
   accumulate: true,
+};
+
+export const titleResource = {
+  ...baseManifest,
+  path: TITLE_API,
 };