+++ CHANGELOG.md
@@ -4,5 +4,6 @@
 
 ### Stories
 * [UIREC-28](https://issues.folio.org/browse/UIREC-28) Add Title for receiving/check-in
+* [UIREC-27](https://issues.folio.org/browse/UIREC-27) View title in receiving area
 * [UIREC-26](https://issues.folio.org/browse/UIREC-26) Create three pane layout for receiving
 * New app created with stripes-cli